feat: Add categoryOrder to social media links
This commit adds a new property called categoryOrder to the social media links in order to define the order in which the links should be displayed. The categoryOrder is an object that maps each category to a number, allowing for custom sorting of the links. This change ensures that the social media links are displayed in a consistent and logical order.